\name{R2DOCX-package}
\alias{R2DOCX-package}
\alias{R2DOCX}
\docType{package}
\title{
Microsoft Word document generation from R
}
\description{
Enable R to add content into a Word 2007 document - available elements are:
\enumerate{
  \item{plots}
  \item{tables}
  \item{paragraphs}
  \item{Tables of contents}
}
}
\details{
\tabular{ll}{
Package: \tab R2DOCX\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-06-06\cr
License: \tab GPL-3
}

	\cr
	Methods for adding content :\cr
		\enumerate{
			\item{\code{addPlot} for adding plots}
			\item{\code{addParagraph} for adding paragraphs}
			\item{\code{addTable} for adding tables}
			\item{\code{addTOC} for adding tables of contents}
			\item{\code{replaceText} for text replacement}
		}
}
\author{
Author: David Gohel <\email{david.gohel@gmail.com}; \email{david.gohel@lysis-consultants.fr}>\cr
Maintainer: David Gohel <\email{david.gohel@gmail.com}; \email{david.gohel@lysis-consultants.fr}>\cr
}
\seealso{
\code{\linkS4class{Docx}}
}
\keyword{ package }

